chore(conversation): filename convention + folder split for message and message list features
