+++ conversation/frontend/package.json.template
@@ -29,7 +29,8 @@
   },
   "husky": {
     "hooks": {
-      "pre-commit": "lint-staged"
+      "pre-commit": "lint-staged",
+      "pre-push": "pnpm test"
     }
   },
   "lint-staged": {
