--- conversation/frontend/src/features/Menu/components/DesktopMenu.css
@@ -1,3 +1,0 @@
-.user-folder {
-    margin: -6px 0;
-}
