+++ conversation/frontend/src/features/index.ts
@@ -1,4 +1,4 @@
-export * from './Modals/CreateFolderModal';
-export * from './Modals/RenameFolderModal';
-export * from './Modals/TrashFolderModal';
-export * from './Menu';
+export * from './menu';
+export * from './modals/CreateFolderModal';
+export * from './modals/RenameFolderModal';
+export * from './modals/TrashFolderModal';
