+++ conversation/frontend/src/features/message-list/message-preview/components/SenderAvatar.tsx
