+++ conversation/frontend/src/features/message-list/message-preview/index.tsx
@@ -3,10 +3,10 @@ import { IconPaperclip, IconUndo } from '@edifice.io/react/icons';
 import { useTranslation } from 'react-i18next';
 import { useParams } from 'react-router-dom';
 import { MessageMetadata } from '~/models';
+import { useMessageUserDisplayName } from '../../../hooks/useUserDisplayName';
 import RecipientAvatar from './components/RecipientAvatar';
 import { RecipientListPreview } from './components/RecipientListPreview';
 import { SenderAvatar } from './components/SenderAvatar';
-import { useMessageUserDisplayName } from './hooks/useUserDisplayName';
 
 export interface MessagePreviewProps {
   message: MessageMetadata;
