+++ conversation/frontend/src/hooks/useRecipientAvatar.ts
